move _get_entries_for_section to tests/test_refine.py; as it's used only in tests now
